hospital-frontend/requirements.txt
# Django Hospital Management Frontend
# Requirements for Django frontend connecting to API Gateway v2.2.0

# Core Django
Django==4.2.7
djangorestframework==3.14.0

# API Integration
requests==2.31.0
httpx==0.25.2

# WebSocket Support for Real-time Features
channels==4.0.0
channels-redis==4.1.0
websockets==12.0

# Authentication & Security
django-cors-headers==4.3.1
PyJWT==2.8.0
cryptography>=42.0.0

# Forms & UI
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-widget-tweaks==1.5.0

# Database (for session storage, not main data)
psycopg2-binary==2.9.9

# Static Files & Media
whitenoise==6.6.0
Pillow>=10.0.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Production
gunicorn==21.2.0
python-decouple==3.8

# Date/Time handling
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Pagination
django-pure-pagination==0.3.0

# Charts & Analytics
plotly==5.17.0
pandas==2.1.4

# Caching
django-redis==5.4.0

# Environment
python-dotenv==1.0.0
